flask_ocr/requirements.txt
Flask
pillow
pytesseract
pdf2image
python-docx
